+++ bioneuron_oracle/conftest.py
@@ -1,1 +1,4 @@
 from nengo.conftest import *  # noqa: F401, F403
+
+import seaborn as sns
+sns.set(context='poster')
